+++ driver/src/main.rs
@@ -1,13 +1,11 @@
 //! Server side entry point
 //! This to be conversted into server
 
-use frtb_engine::prelude::*;
-use std::{process, str::FromStr};
+use base_engine::prelude::*;
+use std::process;
 //use pretty_env_logger;
-use log::{debug, info, trace, warn};
+use log::info;
 use serde::{Serialize, Deserialize};
-use serde_json::value::Serializer;
-use chrono::NaiveDate;
 
 fn main() {
     dotenv::dotenv().ok();
@@ -18,10 +16,6 @@ fn main() {
 
     let data = conf.build_data();
 
-    //let a = FRTBFilter::On(vec![("Anatoly".to_string(), "Bugakov".to_string())]);
-    //let z = serde_json::to_string::<FRTBFilter>(&a).unwrap();
-    //debug!("FRTBFilter: {:?}", z);
-
     let message: Message = serde_json::from_str(JSON3).unwrap();
 
     //recoverable. If not valid use default
@@ -29,13 +23,12 @@ fn main() {
     .unwrap(); // recoverable, but for now ok to unwrap()
 
     // Example setup to validate possible filter/groupby
-    let groups: Vec<String> = vec!["TradeId".to_string(), "RiskClass".to_string(), "RiskFactor".to_string()];
-    let groups = data.derive_groups(groups);
+    //let groups = data.derive_groups(groups);
     //println!("Groups: {:?}", groups);
 
     match message {
         Message::Request{ params: conf, ..} => {
-            match frtb_engine::sa_capital(conf, &data, default_params){
+            match base_engine::sa_capital(conf, &data, default_params){
                 Err(e) =>{ // eventually will be tokio::spawn_blocking
                     eprintln!("Application error: {:#?}", e);
                     process::exit(1);
@@ -61,6 +54,7 @@ enum Message {
 #[derive(Serialize, Deserialize)]
 struct PlaceHolder(u8);
 
+/*
 /// Sample request
 const JSON: &str = r#"
 {"type": "Request",
@@ -88,7 +82,7 @@ const JSON2: &str = r#"
         "filters": [{"In":[["LegalEntity", ["EMEA"]], ["Country", ["UK", "China"]]]}]
     }
 }"#;
-
+*/
 /// Sample request 2
 const JSON3: &str = r#"
 {"type": "Request",
